Write func rework

+++ lib/qhg_nn_sequential_sink.h
@@ -6,7 +6,7 @@
 void qhg_nn_sequential_sink_d(qhg_spinor_field [], qhg_spinor_field [], int, qhg_thrp_nn_sink_params);
 void qhg_nn_sequential_sink_u(qhg_spinor_field [], qhg_spinor_field [], qhg_spinor_field [], int, qhg_thrp_nn_sink_params);
 void qhg_mesons_sequential_sink(qhg_spinor_field [], qhg_spinor_field [], int, qhg_thrp_nn_sink_params);
-void qhg_phase_sequential_sink(qhg_spinor_field [], qhg_spinor_field [], int mom[3], int source_coords[4], int);
+void qhg_phase_sequential_sink(qhg_spinor_field [], qhg_spinor_field [], int mom[3], int source_coords[4], double);
 
   
 #endif /* _QHG_NN_SEQUENTIAL_SINKS_H */
